+++ src/main/java/grit/stockIt/domain/mission/service/MissionService.java
@@ -15,16 +15,11 @@ import grit.stockIt.domain.mission.enums.MissionTrack;
 import grit.stockIt.domain.mission.enums.MissionType;
 import grit.stockIt.domain.mission.repository.MissionProgressRepository;
 import grit.stockIt.domain.mission.repository.MissionRepository;
-import grit.stockIt.domain.notification.event.MissionCompletedEvent;
 import grit.stockIt.domain.order.entity.OrderMethod;
 import grit.stockIt.domain.order.event.TradeCompletionEvent;
-import grit.stockIt.domain.title.entity.MemberTitle;
-import grit.stockIt.domain.title.entity.Title;
-import grit.stockIt.domain.title.repository.MemberTitleRepository;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import grit.stockIt.domain.stock.entity.Stock;
@@ -47,17 +42,15 @@ public class MissionService {
     private final MemberRepository memberRepository;
     private final MissionRepository missionRepository;
     private final MissionProgressRepository missionProgressRepository;
-    private final MemberTitleRepository memberTitleRepository;
     private final AccountRepository accountRepository;
     private final AccountStockRepository accountStockRepository; // [추가됨] 홀딩 여부 확인용
     private final StockRepository stockRepository;
-    private final ApplicationEventPublisher eventPublisher;
     private final MissionConditionEvaluator missionConditionEvaluator;
     private final MissionProgressCalculator missionProgressCalculator;
     private final MissionTrackPolicy missionTrackPolicy;
+    private final MissionRewardService missionRewardService;
 
     private static final long JUNK_STOCK_MARKET_CAP_THRESHOLD = 100000000000L;
-    private static final int MISSION_COMPLETION_ACTIVITY_POINTS = 10;
     /**
      * [1] (이벤트 수신) 거래 이벤트 발생 시 미션 진행도 업데이트
      * - 일반 미션 갱신 로직
@@ -149,25 +142,10 @@ public class MissionService {
                             member.getName(), profitInt, newTotalProfit);
 
                     // Legend 달성 체크 (현재 총 수익금 기반으로 점수 환산하여 체크)
-                    checkLegendTier(member, getActivityScore(member) + missionProgressCalculator.calculateScoreFromProfit(newTotalProfit));
+                    missionRewardService.checkLegendTier(member, getActivityScore(member) + missionProgressCalculator.calculateScoreFromProfit(newTotalProfit));
                 });
     }
 
-    // 레전드 미션(903) 달성 체크
-    private void checkLegendTier(Member member, int totalScore) {
-        if (totalScore >= 3600) { // Legend 기준 점수
-            missionRepository.findAllByTrackAndConditionType(MissionTrack.ACHIEVEMENT, MissionConditionType.REACH_LEGEND)
-                    .stream().findFirst().ifPresent(legendMission -> {
-                        missionProgressRepository.findByMemberAndMission(member, legendMission)
-                                .ifPresent(progress -> {
-                                    if (!progress.isCompleted()) {
-                                        progress.setCurrentValue(1);
-                                        checkMissionCompletion(progress); // 칭호 및 1억 지급
-                                    }
-                                });
-                    });
-        }
-    }
     /**
      * [Helper] 현재 활동 점수(Activity Score) 조회
      * - 트래커 미션(998번)의 현재 값을 가져옴
@@ -224,7 +202,7 @@ public class MissionService {
                 .ifPresent(progress -> {
                     if (!progress.isCompleted()) {
                         progress.setCurrentValue(value);
-                        checkMissionCompletion(progress);
+                        missionRewardService.checkMissionCompletion(progress);
                     }
                 });
     }
@@ -246,7 +224,7 @@ public class MissionService {
                 progress.incrementProgress(1);
                 log.info("홀딩 미션 +1일 증가: MemberId={}, MissionId={}, NewValue={}",
                         member.getMemberId(), progress.getMission().getId(), progress.getCurrentValue());
-                checkMissionCompletion(progress);
+                missionRewardService.checkMissionCompletion(progress);
             }
         }
     }
@@ -263,7 +241,7 @@ public class MissionService {
                 progress.incrementProgress(valueToIncrease);
                 log.info("미션(누적) 갱신: MissionId={}, Added={}, Current={}",
                         mission.getId(), valueToIncrease, progress.getCurrentValue());
-                checkMissionCompletion(progress);
+                missionRewardService.checkMissionCompletion(progress);
             }
         }
         // B. 달성형 (임계값 돌파 / 최고 기록 갱신) - [수정됨]
@@ -281,7 +259,7 @@ public class MissionService {
 
                 // 목표 달성 여부 체크
                 if (eventValue >= goal) {
-                    checkMissionCompletion(progress);
+                    missionRewardService.checkMissionCompletion(progress);
                 }
             }
         }
@@ -308,80 +286,6 @@ public class MissionService {
 
         log.info("총 {}건의 연속 출석 기록이 일괄 초기화되었습니다.", updatedCount);
     }
-
-    // --- [공통 로직] 완료 처리, 보상, 초기화 ---
-
-    public void checkMissionCompletion(MissionProgress progress) {
-        if (progress.getStatus() == MissionStatus.COMPLETED || !progress.isCompleted()) {
-            return;
-        }
-
-        progress.complete();
-        log.info("미션 완료: MemberId={}, MissionId={}", progress.getMember().getMemberId(), progress.getMission().getId());
-
-        Mission mission = progress.getMission();
-        Reward reward = mission.getReward();
-        
-        // 보상 지급
-        distributeReward(progress.getMember(), reward);
-        
-        // 출석 미션은 알림 발송 제외
-        if (mission.getConditionType() != MissionConditionType.LOGIN_COUNT) {
-            // 미션 완료 알림 이벤트 발행
-            publishMissionCompletedEvent(progress.getMember(), mission, reward);
-        }
-        
-        activateNextMission(progress);
-        handleMissionChain(progress);
-        checkSeedCopierAchievement(progress.getMember());
-
-        // [추가] 3. 활동 점수(Activity Score) 반영
-        updateActivityScore(progress.getMember());
-    }
-    /**
-     * [신규] 활동 점수 업데이트
-     * - 미션 1개 완료 시 +10점 (예시)
-     * - 최대 점수(GoalValue)를 넘을 수 없음
-     */
-    private void updateActivityScore(Member member) {
-        // 활동 점수 트래커 조회 (ID 998 or Type=ACTIVITY_SCORE)
-        missionProgressRepository.findByMemberAndMissionTypeWithMission(member, MissionTrack.ACHIEVEMENT, MissionConditionType.ACTIVITY_SCORE)
-                .ifPresent(tracker -> {
-                    int maxScore = tracker.getMission().getGoalValue(); // 예: 1000점
-                    int currentScore = tracker.getCurrentValue();
-
-                    if (currentScore < maxScore) {
-                        int pointsToAdd = MISSION_COMPLETION_ACTIVITY_POINTS; // 미션 당 점수 (기획에 따라 조정)
-                        int newScore = Math.min(currentScore + pointsToAdd, maxScore);
-                        tracker.setCurrentValue(newScore);
-                        log.info("활동 점수 획득: Member={}, Current={}, Max={}", member.getName(), newScore, maxScore);
-                    }
-                });
-    }
-    /**
-     * [신규] 시드 복사기 (누적 미션 30회) 체크
-     */
-    private void checkSeedCopierAchievement(Member member) {
-        // 완료된 미션 총 개수 조회
-        long completedCount = missionProgressRepository.countByMemberAndStatus(member, MissionStatus.COMPLETED);
-
-        missionProgressRepository.findByMemberAndMissionTypeWithMission(member, MissionTrack.ACHIEVEMENT, MissionConditionType.TOTAL_MISSION_COUNT)
-                .ifPresent(progress -> {
-                    if (!progress.isCompleted()) {
-                        // 현재 완료 횟수를 진행도에 반영
-                        progress.setCurrentValue((int) completedCount);
-                        if (completedCount >= progress.getMission().getGoalValue()) { // 30회
-                            // 재귀 호출 방지를 위해 직접 complete 호출 (혹은 checkMissionCompletion 호출)
-                            // 여기서는 간단히 내부 로직 실행
-                            progress.complete();
-                            distributeReward(member, progress.getMission().getReward());
-                            log.info("'시드 복사기' 업적 달성! 총 완료 미션: {}개", completedCount);
-                        }
-                    }
-                });
-    }
-
-
 
     /**
      * [신규] 인생 2회차 (파산 신청) API 로직
@@ -423,7 +327,7 @@ public class MissionService {
 
         bankruptcyProgress.setCurrentValue(bankruptcyProgress.getMission().getGoalValue()); // 조건 충족 표시
         bankruptcyProgress.complete();
-        distributeReward(member, bankruptcyProgress.getMission().getReward());
+        missionRewardService.distributeReward(member, bankruptcyProgress.getMission().getReward());
 
         log.info("파산 신청 승인! 구조지원금 지급 완료. Member={}", member.getName());
 
@@ -436,56 +340,6 @@ public class MissionService {
 
         log.info("파산 승인 및 티어 초기화 완료: Member={}", member.getName());
         return bankruptcyProgress.getMission().getReward();
-    }
-
-    private void distributeReward(Member member, Reward reward) {
-        if (reward == null) return;
-
-        if (reward.getMoneyAmount() > 0) {
-            accountRepository.findByMemberAndIsDefaultTrue(member)
-                    .ifPresentOrElse(
-                            acc -> {
-                                acc.increaseCash(BigDecimal.valueOf(reward.getMoneyAmount()));
-                                log.info("보상 지급: {}원", reward.getMoneyAmount());
-                            },
-                            () -> log.error("보상 지급 실패: 기본 계좌 없음 MemberId={}", member.getMemberId())
-                    );
-        }
-
-        if (reward.getTitleToGrant() != null) {
-            if (!memberTitleRepository.existsByMemberAndTitle(member, reward.getTitleToGrant())) {
-                member.addMemberTitle(MemberTitle.builder()
-                        .member(member)
-                        .title(reward.getTitleToGrant())
-                        .build());
-                log.info("칭호 지급: {}", reward.getTitleToGrant().getName());
-            }
-        }
-    }
-
-    /**
-     * 미션 완료 알림 이벤트 발행
-     */
-    private void publishMissionCompletedEvent(Member member, Mission mission, Reward reward) {
-        Long rewardId = reward != null ? reward.getId() : null;
-        Long moneyAmount = reward != null ? reward.getMoneyAmount() : 0L;
-        String titleName = (reward != null && reward.getTitleToGrant() != null) 
-                ? reward.getTitleToGrant().getName() 
-                : null;
-
-        MissionCompletedEvent event = new MissionCompletedEvent(
-                member.getMemberId(),
-                mission.getId(),
-                mission.getName(),
-                mission.getTrack(),
-                rewardId,
-                moneyAmount,
-                titleName
-        );
-
-        eventPublisher.publishEvent(event);
-        log.debug("미션 완료 이벤트 발행: memberId={}, missionId={}, missionName={}", 
-                member.getMemberId(), mission.getId(), mission.getName());
     }
 
     /**
@@ -524,52 +378,7 @@ public class MissionService {
             if (!progress.isCompleted()) {
                 log.info("랭커 등극! 칭호 지급: MemberId={}", memberId);
                 progress.setCurrentValue(10); // 목표치(10) 달성 처리
-                checkMissionCompletion(progress); // 보상(칭호) 지급 및 완료 처리
-            }
-        }
-    }
-
-    private void activateNextMission(MissionProgress completedProgress) {
-        Mission completedMission = completedProgress.getMission();
-        Member member = completedProgress.getMember();
-
-        if (completedMission.getTrack() == MissionTrack.DAILY || completedMission.getTrack() == MissionTrack.ACHIEVEMENT) {
-            return;
-        }
-
-        Mission nextMission = completedMission.getNextMission();
-
-        if (nextMission != null) {
-            log.info("다음 미션 활성화: MissionId={}", nextMission.getId());
-            missionProgressRepository.findByMemberAndMission(member, nextMission)
-                    .orElseGet(() -> {
-                        MissionProgress newProgress = MissionProgress.builder()
-                                .member(member)
-                                .mission(nextMission)
-                                .status(MissionStatus.INACTIVE)
-                                .build();
-                        member.addMissionProgress(newProgress);
-                        return newProgress;
-                    }).activate();
-        } else if (completedMission.getType() == MissionType.ADVANCED) {
-            log.info("트랙 최종 완료: Track={}", completedMission.getTrack());
-            resetMissionTrack(member, completedMission.getTrack());
-        }
-    }
-
-    public void resetMissionTrack(Member member, MissionTrack track) {
-        log.info("트랙 초기화 시작: MemberId={}, Track={}", member.getMemberId(), track);
-        List<MissionProgress> progressList = missionProgressRepository.findAllByMemberAndMission_Track(member, track);
-
-        for (MissionProgress progress : progressList) {
-            progress.reset();
-            progress.deactivate();
-
-            // 트랙의 첫 번째 미션(중급 1단계)만 다시 활성화
-            if (progress.getMission().getType() == MissionType.INTERMEDIATE
-                    && missionTrackPolicy.isFirstMissionInTrack(progress.getMission().getId())) {
-                progress.activate();
-                log.info("트랙 첫 미션 재활성화: MissionId={}", progress.getMission().getId());
+                missionRewardService.checkMissionCompletion(progress); // 보상(칭호) 지급 및 완료 처리
             }
         }
     }
@@ -653,7 +462,7 @@ public class MissionService {
         }
 
         attendanceProgress.incrementProgress(1);
-        checkMissionCompletion(attendanceProgress);
+        missionRewardService.checkMissionCompletion(attendanceProgress);
         return attendanceProgress.getMission().getReward();
     }
 
@@ -674,7 +483,7 @@ public class MissionService {
                     if (progress.getStatus() != MissionStatus.COMPLETED && !progress.isCompleted()) {
                         progress.incrementProgress(1);
                         log.info("일일 미션({}) 진행도 갱신: MemberId={}", type, member.getMemberId());
-                        checkMissionCompletion(progress);
+                        missionRewardService.checkMissionCompletion(progress);
                     }
                 });
     }
@@ -684,50 +493,6 @@ public class MissionService {
     private Member getMemberByEmail(String email) {
         return memberRepository.findByEmail(email)
                 .orElseThrow(() -> new EntityNotFoundException("회원을 찾을 수 없습니다. Email: " + email));
-    }
-
-    private void handleMissionChain(MissionProgress completedProgress) {
-        Member member = completedProgress.getMember();
-        Mission mission = completedProgress.getMission();
-
-        // 일일 출석 완료 시 -> 연속 출석 업적 갱신
-        if (mission.getTrack() == MissionTrack.DAILY &&
-                mission.getConditionType() == MissionConditionType.LOGIN_COUNT) {
-            log.info("연쇄 업적 갱신 시도: 일일 출석 -> 연속 출석");
-            updateSpecificAchievement(member, MissionConditionType.LOGIN_STREAK, 1);
-        }
-    }
-
-    private void updateSpecificAchievement(Member member, MissionConditionType conditionType, int valueToIncrease) {
-        // 1. Optional -> List로 변경하여 해당 타입의 모든 업적 조회 (예: 3일, 7일, 30일 연속 등)
-        List<Mission> achievements = missionRepository
-                .findAllByTrackAndConditionType(MissionTrack.ACHIEVEMENT, conditionType);
-
-        if (achievements.isEmpty()) return;
-
-        // 2. 조회된 모든 업적에 대해 진행도 업데이트 반복
-        for (Mission achievement : achievements) {
-            MissionProgress achievementProgress = missionProgressRepository
-                    .findByMemberAndMission(member, achievement)
-                    .orElseGet(() -> {
-                        MissionProgress newProgress = MissionProgress.builder()
-                                .member(member)
-                                .mission(achievement)
-                                .status(MissionStatus.IN_PROGRESS)
-                                .build();
-                        member.addMissionProgress(newProgress);
-                        return newProgress;
-                    });
-
-            // 이미 완료된 업적은 패스
-            if (achievementProgress.getStatus() != MissionStatus.COMPLETED) {
-                achievementProgress.incrementProgress(valueToIncrease);
-                log.info("업적 미션({}) 갱신: MissionId={}, NewValue={}",
-                        achievement.getName(), achievement.getId(), achievementProgress.getCurrentValue());
-
-                checkMissionCompletion(achievementProgress);
-            }
-        }
     }
 
     // findDailyAttendanceMission 제거 (직접 쿼리 사용으로 대체됨)
